Try to workaround bug #441 - All LXQt::Settings and QSettings file change monitoring stop working.
Detailed descriptions:
Since Qt 5.4, QSettings uses QSaveFile to save the config files.
https://github.com/qtproject/qtbase/commit/8d15068911d7c0ba05732e2796aaa7a90e34a6a1#diff-e691c0405f02f3478f4f50a27bdaecde
QSaveFile will save the content to a new temp file, and replace the old file later.
Hence the existing config file is not changed. Instead, it's deleted and then replaced.
This new behaviour unfortunately breaks QFileSystemWatcher.
After file deletion, we can no longer receive any new change notifications.
However, QFileSystemWatcher does not provide a way for us to know if a file is deleted.
Luckily, I found a workaround: If the file path no longer exists in the watcher's files(), this file is deleted.
This fix workaround the issue by reinstalling the file watcher for the new file if the old file is deleted.