Add support for flatpak
Flatpak sandbox always starts an app with PID 1 or 2. Even if allow an app to own org.kde.StatusNotifierItem-1-1 service, it will conflict with other flatpak'ed apps, since all of them will want to own the same name.

This commit removes owning a service to fix tray icon in flatpak, as KSNI did.