Added an option to show context menus on releasing mouse button
It brings the behavior of MS Windows to Qt apps. The key is `context_menu_on_mouse_release` under the `[General]` section of `~/.config/lxqt/lxqt.conf`, and it can be `true` or `false` (`false` by default).

Pros:

 1. With this option enabled, the user can right click without fearing that a menu-item might be activated by mistake. Many users have experienced unintentional activation.
 2. It makes right click be a real click (= press + release) and consistent with left click.
 3. LXQt might be the first Linux DE providing its users with this option. That could motivate others to take it seriously. In particular, Qt might become more self-consistent about it.

Cons:

 1. Running apps can't see its toggling. App restart — or preferably, session restart — is needed.
 2. There's a hidden problem in `libqtxdg` (inherited from a similar problem in Qt) and another one in lxqt-panel's main menu. But they can be fixed easily.
 3. Of course, nothing can be done about non-Qt apps.

NOTE: It can't be confusing to users because context menus are already shown on mouse release in notification area (an inconsistency not noticed by many).